requirements.txt
psycopg2-binary
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1

matplotlib==3.8.2
reportlab==4.0.9

psycopg2-binary==2.9.9
requests==2.31.0
pytz==2023.3

email-validator 
Flask-Mail
reportlab